Version 3
Much more modular, clean and commented code than before. Also changed
the way the piece are checked if they can be placed on the given cell of
the Board. Added a driver class and a Result class, which will hold the
unique solutions for the given configuration. The driver class also show
the time taken to find all unique solutions in milliseconds.